some refraction, still not working

+++ includes/types.h
@@ -6,7 +6,7 @@
 /*   By: albertvanandel <albertvanandel@student.      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/09/12 16:10:32 by ccaljouw      #+#    #+#                 */
-/*   Updated: 2023/10/12 12:26:24 by cariencaljo   ########   odam.nl         */
+/*   Updated: 2023/10/12 18:31:39 by cariencaljo   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ typedef struct s_pixel
 	float		facing_ratio;
 	float		hit_distance;
 	float		hit_height; // nodig?
-	float		refl_count;
+	int			refl_count;
 	int			color;
 }	t_px;
